backend/requirements.txt
Flask
Flask-Cors
PyYAML
openai
python-dotenv